Rename collection_management folder to collectionManagement
