+++ app/packs/src/components/collectionManagement/MySharedCollections.js
