+++ app/packs/src/components/collectionManagement/SharedWithMeCollections.js
